Commit: Ottimizzazione oscillatori

+++ FractalWindow/Fractal.h
@@ -2,12 +2,10 @@
 #include "threeValuesArray.h"
 #include <SFML/Graphics/Color.hpp>
 #include <SFML/Graphics/Shader.hpp>
-
+//TODO: checking the types
 using namespace sf;
 class Fractal
 {
-	
-
 public:
 	//Fractal();
 
@@ -19,29 +17,29 @@ public:
 	sf::Color paint_fractal(int n, int max_iter);
 	//get methods
 	int getMaxIter() { return max_iter; };
-	int getMinRe() { return min_re; };
-	int getMaxRe() { return max_re; };
-	int getMinIm() { return min_im; };
-	int getMaxIm() { return max_im; };
+	double getMinRe() { return min_re; };
+	double getMaxRe() { return max_re; };
+	double getMinIm() { return min_im; };
+	double getMaxIm() { return max_im; };
 	int getW() { return W; };
 	int getH() { return H; };
 	int getXfp() { return x_fp; };
 	void setXfp(double x_mouse);
 	void setYfp(double y_mouse);
 	int getYfp() { return y_fp; };
-	int getXpoint() { return xPoint; };
+	double getXpoint() { return xPoint; };
 	void setXpoint(double x);
-	int getYpoint() { return yPoint; };
+	double getYpoint() { return yPoint; };
 	void setYpoint(double y);
 
 private:
 	const int max_iter { 128 };     //number of iterations
-	const double min_re { -2 };     //minimum value real axis
-	const double max_re { 2 };      //maximum value real axis
-	const double min_im { -2 };    //minimum value immaginary axis
-	const double max_im { 2 };  
-	const double W{ 600 };
-	const double H{ 600 };
+	const double min_re { -2.0 };     //minimum value real axis
+	const double max_re { 2.0 };      //maximum value real axis
+	const double min_im { -2.0 };    //minimum value immaginary axis
+	const double max_im { 2.0 };  
+	const int W{ 600 };
+	const int H{ 600 };
 	int x_fp {0};
 	int y_fp {0};
 	double xPoint;
